GEOMETRY-45: renaming packages in hull and enclosing modules to avoid duplicating root 'euclidean' package
