+++ commons-geometry-enclosing/src/test/java/org/apache/commons/geometry/enclosing/euclidean/threed/SphereGeneratorTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.geometry.euclidean.threed.enclosing;
+package org.apache.commons.geometry.enclosing.euclidean.threed;
 
 import java.util.ArrayList;
 import java.util.Arrays;
